third_party/WebKit/LayoutTests/external/wpt/html/semantics/interactive-elements/the-menu-element/menuitem-label-expected.txt

This is a testharness.js-based test.
FAIL Menuitem.label should be the label attribute content or the result of stripping and collapsing ASCII whitespace from the child text content. assert_equals: expected (string) "item1" but got (undefined) undefined
Harness: the test ran to completion.
